fix (BowlBall.tsx): Adjusting display name teams visible for mobile dimension
Hello, I noticed that when reducing the screen size for the mobile versions of the project, the names of the teams are not visible when checking the X-Ray checkbox.

So I decided to investigate the bug and found that the font-size logic was switched and checking the value of the selected property instead of the forceVisible property.

So I changed the logic of this font-size. Hope I helped in some way, thanks :)