fix: adjusting bug fetch data below the 2002 season.
Hello, when selecting the elimination round session below the 2002 season, the application crashed because there was no data for the 2000/2001, 2001/2002 seasons, even with the minSeasons property having the 2000 season as its value.

So to solve the problem quickly, while the .json files of the respective seasons are not added, I changed the value of the minSeason property to 2002, which is the season where the .json referring to the year is.

I hope I have helped in some way, and I am available for any questions. Thanks :)